блок с инф. об ответственных и описанием проекта
