+++ components/ProjectPagePublicPath/Contener4CuTuDi&PridjectDescr/AboutProject/AboutProject.module.sass
@@ -4,6 +4,8 @@
   position: relative
   height: 100%
   font-size: 26px
+  // left: 10%
+  // width: 60%
 
   & h1
     text-align: center
