+++ components/ProjectPagePublicPath/Contener4CuTuDi&PridjectDescr/Contener4CuTuDi&PridjectDescr.module.sass
@@ -1,3 +1,5 @@
 .contener
-    display: inline
+    display: flex 
+    //  flex-direction: column
+
