+++ components/ProjectPagePublicPath/Contener4CuTuDi&PridjectDescr/InfoAboutCustomerTutorDirector/InfoAboutCuTuDi.module.sass
@@ -3,6 +3,12 @@
 .InfoAboutCuTuDiInner
   position: relative
   height: 100%
+  margin-right: 78px
+  // margin-top: auto
+  // margin-bottom: auto
+  
+  // min-width: 40%
+
   font-size: 26px
 
   & h1
@@ -11,6 +17,7 @@
     line-height: 1.2em
     margin-bottom: 30px
     font-weight: 600
+    
   .red
     color: #9F2D20
   & p
@@ -19,11 +26,23 @@
   
 
   .contener
-    width: 37%
+    width: 100%
 
+  // .box
+  //     &:nth-child(3n)
+  //       border-color: #c2333300
+  //       background: #c2333333
+
+  //     &:nth-child(3n - 1)
+  //       border-color: #a7c5bd00
+  //       background:#a7c5bd99
+
+  //     &:nth-child(3n - 2)
+  //       border-color: #fcaf3900
+  //       background:#fcaf394d
   .cu
     position: relative
-    width: max-width
+    // width: max-width
     padding: 5px 40px
     border: 5px solid black
     font-size: 1em
@@ -37,17 +56,17 @@
     // border-color: #A7C5BD
     // background: #A7C5BD
     border-radius: 25px
-    &:nth-child(3n)
-        border-color: #c2333300
-        background: #c2333333
+    // &:nth-child(3n)
+    //     border-color: #c2333300
+    //     background: #c2333333
 
-    &:nth-child(3n - 1)
-        border-color: #a7c5bd00
-        background:#a7c5bd99
+    // &:nth-child(3n - 1)
+    //     border-color: #a7c5bd00
+    //     background:#a7c5bd99
 
-    &:nth-child(3n - 2)
-        border-color: #fcaf3900
-        background:#fcaf394d
+    // &:nth-child(3n - 2)
+    //     border-color: #fcaf3900
+    //     background:#fcaf394d
     
   .hidt
     text-align: center
