+++ pages/project_admin/index.tsx
@@ -1,8 +1,8 @@
 import Head from "next/head";
 import Header from "../../components/Header/";
 import TagSliderForCurProject from "../../components/ProjectPagePublicPath/TagSliderForCurProject";
-import AboutProject from "../../components/ProjectPagePublicPath/AboutProject/";
-import InfoAboutCuTuDi from "../../components/ProjectPagePublicPath/InfoAboutCustomerTutorDirector";
+import AboutProject from "../../components/ProjectPagePublicPath/Contener4CuTuDi&PridjectDescr/AboutProject";
+import InfoAboutCuTuDi from "../../components/ProjectPagePublicPath/Contener4CuTuDi&PridjectDescr/InfoAboutCustomerTutorDirector";
 import Container from "../../components/Container";
 import Footer from "../../components/Footer";
 import Spacer from "../../components/Spacer";
